rename cuda to gpu
